fix(31-08): pin AutoDeltaPackedMap/Set/Vector wire counts to uint32_t (BITS-03)
Mirror the Task-1 AutoDeltaMap fix across the rest of the AutoDelta* family so
every serialized count binds the existing 4-byte Archive overload on LLP64 x64
(raw size_t = 8 bytes = overload-resolution FAILURE; the 31-05 defect class).
Wire byte-identical to the 32-bit server; element put/get byte-unchanged.

- AutoDeltaSet.h + AutoDeltaVector.h: m_baselineCommandCount / baselineCommandCount
  MEMBER size_t -> uint32_t; pack/packDelta size()/commands.size()/v.size() puts
  guarded with assert(<= UINT32_MAX) + static_cast<uint32_t>; static_cast<size_t>(0)
  placeholder -> uint32_t; unpack/unpackDelta count locals (commandCount, bcc,
  targetBaselineCommandCount, skipCount, loop indices) -> uint32_t so the skip
  arithmetic stays type-consistent (modular-identical to the 32-bit build)
- AutoDeltaPackedMap.h: derives from the (already-fixed) AutoDeltaMap; its own
  internal_pack/internal_unpack count locals -> uint32_t + the placeholder ->
  uint32_t (countCharacter() returns int -> already binds the 4-byte signed
  overload, left as-is). Command::index stays unsigned short by design (a 16-bit
  net-traffic optimization, not a count).
- #include <cassert>/<cstdint> in each
- the include/Archive/AutoDelta*.h copies are thin forwarders to these src/shared
  headers, so both consumer include paths are fixed by these edits
- verified: -Scope autodelta sweep (10 TUs incl. Server/SharedGroupObjectTemplate
  + UnicodeAutoDeltaPackedMap + the scratch instantiation TU) compiles exit 0,
  0 C2665/C2668 (was 8 from AutoDeltaSet.h + the AutoDeltaMap cascade)

+++ src/external/ours/library/archive/src/shared/AutoDeltaPackedMap.h
@@ -11,6 +11,8 @@
 // ======================================================================
 
 #include "Archive/AutoDeltaMap.h"
+#include <cassert>
+#include <cstdint>
 #include <stdio.h>
 
 // ======================================================================
@@ -99,8 +101,8 @@ namespace Archive
 		char temp[200];
 		typename AutoDeltaMap<KeyType, ValueType, ObjectType>::Command c;
 
-		Archive::put(target, countCharacter(buffer,':'));
-		Archive::put(target, static_cast<size_t>(0)); // baselineCommandCount
+		Archive::put(target, countCharacter(buffer,':')); // int -> 4-byte signed-int overload
+		Archive::put(target, static_cast<uint32_t>(0)); // baselineCommandCount
 		
 		int tempPos = 0;
 		for (std::string::const_iterator i=buffer.begin(); i!=buffer.end(); ++i)
@@ -127,8 +129,8 @@ namespace Archive
 		char temp[200];
 		
 		typename AutoDeltaMap<KeyType, ValueType, ObjectType>::Command c;
-		size_t commandCount;
-		size_t baselineCommandCount;
+		uint32_t commandCount;
+		uint32_t baselineCommandCount;
 
 		Archive::get(source, commandCount);
 		Archive::get(source, baselineCommandCount);
@@ -139,7 +141,7 @@ namespace Archive
 		}
 		else
 		{
-			for (size_t i = 0; i < commandCount; ++i)
+			for (uint32_t i = 0; i < commandCount; ++i)
 			{
 				Archive::get(source, c.cmd);
 				Archive::get(source, c.key);
